# credentials/alts: Optimize reads