--- a/credentials/alts/internal/conn/record.go
+++ b/credentials/alts/internal/conn/record.go
@@ -87,13 +87,8 @@ type conn struct {
 	net.Conn
 	crypto ALTSRecordCrypto
 	// buf holds data that has been read from the connection and decrypted,
-	// but has not yet been returned by Read.
-	buf []byte
-	// bufPointer holds the entire decrypted record, even bytes that have
-	// been returned by read. It is used to restore buf to it's initial
-	// capacity after each frame is decrypted. It is also used to return the
-	// buffer to the buffer pool.
-	bufPointer         *[]byte
+	// but has not yet been returned by Read. It is a sub-slice of protected.
+	buf                []byte
 	payloadLengthLimit int
 	// protected holds data read from the network but have not yet been
 	// decrypted. This data might not compose a complete frame.
@@ -128,7 +123,9 @@ func NewConn(c net.Conn, side core.Side, recordProtocol string, key []byte, prot
 	// We increase the size of the buffer by the required amount if can't hold a
 	// complete encrypted record.
 	protectedPointer := mem.DefaultBufferPool().Get(max(altsReadBufferInitialSize, len(protected)))
-	protectedBuf := (*protectedPointer)[:copy(*protectedPointer, protected)]
+	// Copy additional data from hanshaker service.
+	copy(*protectedPointer, protected)
+	protectedBuf := (*protectedPointer)[:len(protected)]
 
 	altsConn := &conn{
 		Conn:               c,
@@ -139,7 +136,6 @@ func NewConn(c net.Conn, side core.Side, recordProtocol string, key []byte, prot
 		writeBuf:           make([]byte, altsWriteBufferInitialSize),
 		nextFrame:          protectedBuf,
 		overhead:           overhead,
-		bufPointer:         mem.DefaultBufferPool().Get(altsReadBufferInitialSize),
 	}
 	return altsConn, nil
 }
@@ -147,8 +143,8 @@ func NewConn(c net.Conn, side core.Side, recordProtocol string, key []byte, prot
 func (p *conn) Close() error {
 	if !p.isClosed {
 		p.isClosed = true
+		p.buf = nil
 		mem.DefaultBufferPool().Put(p.protectedPointer)
-		mem.DefaultBufferPool().Put(p.bufPointer)
 	}
 	return p.Conn.Close()
 }
@@ -170,7 +166,8 @@ func (p *conn) Read(b []byte) (n int, err error) {
 		// Check whether the next frame to be decrypted has been
 		// completely received yet.
 		if len(framedMsg) == 0 {
-			p.protected = p.protected[:copy(p.protected, p.nextFrame)]
+			copy(p.protected, p.nextFrame)
+			p.protected = p.protected[:len(p.nextFrame)]
 			// Always copy next incomplete frame to the beginning of
 			// the protected buffer and reset nextFrame to it.
 			p.nextFrame = p.protected
@@ -180,18 +177,18 @@ func (p *conn) Read(b []byte) (n int, err error) {
 			if len(p.protected) == cap(p.protected) {
 				// We can parse the length header to know exactly how large
 				// the buffer needs to be to hold the entire frame.
-				length, didParse := ParseMessageLength(p.protected)
+				length, didParse := parseMessageLength(p.protected)
 				if !didParse {
 					// The protected buffer is initialized with a capacity of
 					// larger than 4B. It should always hold the message length
 					// header.
 					panic(fmt.Sprintf("protected buffer length shorter than expected: %d vs %d", len(p.protected), MsgLenFieldSize))
 				}
-				tmp := mem.DefaultBufferPool().Get(int(length))
-				copy(*tmp, p.protected)
-				p.protected = (*tmp)[:len(p.protected)]
+				newProtectedBuf := mem.DefaultBufferPool().Get(int(length))
+				copy(*newProtectedBuf, p.protected)
+				p.protected = (*newProtectedBuf)[:len(p.protected)]
 				mem.DefaultBufferPool().Put(p.protectedPointer)
-				p.protectedPointer = tmp
+				p.protectedPointer = newProtectedBuf
 			}
 			n, err = p.Conn.Read(p.protected[len(p.protected):cap(p.protected)])
 			if err != nil {
@@ -212,27 +209,26 @@ func (p *conn) Read(b []byte) (n int, err error) {
 		}
 		ciphertext := msg[msgTypeFieldSize:]
 
-		// Decrypt directly into the buffer, avoiding a copy to p.buf if
+		// Decrypt directly into the buffer, avoiding a copy from p.buf if
 		// possible.
-		if cap(b) >= len(msg) {
+		if cap(b) >= len(ciphertext) {
 			dec, err := p.crypto.Decrypt(b[:0], ciphertext)
 			if err != nil {
 				return 0, err
 			}
 			return len(dec), nil
 		}
-		// Resize the read buffer if needed to hold the entire decrypted
-		// frame.
-		if cap(*p.bufPointer) < len(ciphertext) {
-			mem.DefaultBufferPool().Put(p.bufPointer)
-			p.bufPointer = mem.DefaultBufferPool().Get(len(ciphertext))
-		}
-		p.buf = *p.bufPointer
-		dec, err := p.crypto.Decrypt(p.buf[:0], ciphertext)
+		// Decrypt requires that if the dst and ciphertext alias, they
+		// must alias exactly. Code here used to use msg[:0], but msg
+		// starts MsgLenFieldSize+msgTypeFieldSize bytes earlier than
+		// ciphertext, so they alias inexactly. Using ciphertext[:0]
+		// arranges the appropriate aliasing without needing to copy
+		// ciphertext or use a separate destination buffer. For more info
+		// check: https://golang.org/pkg/crypto/cipher/#AEAD.
+		p.buf, err = p.crypto.Decrypt(ciphertext[:0], ciphertext)
 		if err != nil {
 			return 0, err
 		}
-		p.buf = p.buf[:len(dec)]
 	}
 
 	n = copy(b, p.buf)