## Background

Internal issue: b/400312873

While benchmarking Google Cloud Storage read performance, the flame graphs showed significant time being spent in `google.golang.org/grpc/credentials/alts/internal/conn.(*conn).Read`.  These changes are done to reduce the time spent in this function. The following optimizations are done:

1. In the handshaker, allocate a buffer only once, outside the `for` loop.
2. Use 32KB (previously 4KB) buffers to read from the network.
3. When a complete ALTS frame doesn't fit into the buffer, directly expand the buffer to the required capacity instead of incrementing in steps of 4KB.
4. If the buffer passed to `Read()` is large enough, use it to store the decrypted message, avoiding an extra copy to `buf`. The default buffer size used by gRPC is 16KB, which also happens to be the size of ALTS records from GCS. This means we usually avoid the extra copy.

## Benchmarks

branch: master
```
go test ./credentials/alts/internal/conn -timeout=2s -bench="Bench"
goos: linux
goarch: amd64
pkg: google.golang.org/grpc/credentials/alts/internal/conn
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
BenchmarkLargeMessage-48              13          90139719 ns/op
PASS
ok      google.golang.org/grpc/credentials/alts/internal/conn   2.316s
```

branch: perfromance (this PR)
```
goos: linux
goarch: amd64
pkg: google.golang.org/grpc/credentials/alts/internal/conn
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
BenchmarkLargeMessage-48              14          80422425 ns/op
PASS
ok      google.golang.org/grpc/credentials/alts/internal/conn   2.139s
```

RELEASE NOTES:
* credentials/alts: Improve read performance by optimizing buffer copies and allocations.