@arjan-bal 2025-04-03
The CI has caught a data race where gRPC is calling `conn.Read` and `conn.Close` concurrently: https://github.com/grpc/grpc-go/actions/runs/14238675443/job/39903307175?pr=8204

There are a few solutions to fix the race:
1. Don't add the `protectedBuf` back to the buffer pool. It will be garbage collected like before.
2. Add synchronization in the http client to serialize calls Read and Close.
3. Add synchronization inside `alts.conn` to serialize Read, Write and Close.

I'll try to see if I can make the fix for 2. If that's not possible, 1 seems fine too.