@arjan-bal 2025-04-07
> Is this ready for another review pass, and are the benchmarks in the description up to date with the change? Or are you still experimenting with doing it without the buffer pools?

Decided to remove the buffer pool after discussion with Doug. I re-ran the benchmark and the performance seems to be very slightly better after removing the buffer pool. The PR description is updated.

> Doug raised a concern about the buffer only growing but not shrinking, which may lead to higher memory usage if the peers sent large ALTS records in the past and there are a large number of ALTS conns. I'm checking what other languages do.

Decided not to solve the buffer shrinking problem since the current implementation already suffers from this and it's not something introduced in this PR.

From discussions with other language maintainers Java has a scatter gather to read multiple buffers from the socket. The list of buffers is being copied to a contiguous array for decryption, so it's not zero copy. I tried implementing a similar [scatter gather style in Go](https://github.com/grpc/grpc-go/compare/master...arjan-bal:alts-scatter-gather?expand=1) and the performance was ~10% worse than the existing implementation on the master branch.